Switch patchpoint identity to <func version, IL offset>
This PR changes the patchpoint cache identity to be a pair <function
version, IL offset> instead of by IP of the return address. This is in
preparation for runtime async changes where suspension inside OSR
methods have to be resumed via resumption stub -> tier 0 -> OSR method.
The latter step needs to retransition into the right OSR method from the
tier 0 resumption code. With the current identity this is problematic
because the return address is not going to be the same as the original
patchpoint that induced the OSR transition.