Interpreter EH support in the runtime
This change adds support for exception handling for cases when
interpreter frames participate in the process. That means when an
exception is thrown from an interpreter frame or when it is propagated
over interpreter frames. This doesn't add all the exception handling
support to the interpreter itself, there is a follow up PR that contains
that code.