[android] Add -keepnativesymbols to Windows build
This is so the symbols are kept in the native libs so that you can
debug straight away in android studio. The cmake that gets generated
for each test app also needed CMAKE_BUILD_TYPE to be set otherwise
the breakpoints would not be hit. It'll just work on non-windows.