[release/9.0] Fix edge cases in Tarjan GC bridge (Android) (#114391)
* Fix dump_processor_state debug code to compile and work on Android (#112970)

Fix typo in GC bridge comparison message (SCCS -> XREFS)

* [mono] Add a few bridge tests (#113703)

* [mono][sgen] Fix DUMP_GRAPH debug option build for tarjan bridge

* [mono][sgen] Don't create ScanData* during debug dumping of SCCs

It serves no purpose and it would later crash the runtime since we didn't patch the lockword back in place.

* [mono][sgen] Fix some null deref crashes in DUMP_GRAPH debug option

* [mono][tests] Add bridge tests

These are ported from some of the bridge tests we had on mono/mono. In order to test them we compare between the output of the new and the tarjan bridge.

* Fix an edge case in the Tarjan GC bridge that leads to losing xref information (#112825)

* Fix an edge case in the Tarjan SCC that lead to losing xref information

In the Tarjan SCC bridge processing there's a color graph used to find out
connections between SCCs. There was a rare case which only manifested when
a cycle in the object graph points to another cycle that points to a bridge
object. We only recognized direct bridge pointers but not pointers to other
non-bridge SCCs that in turn point to bridges and where we already calculated
the xrefs. These xrefs were then lost.

* Add test case to sgen-bridge-pathologies and add an assert to catch the original bug

* Add review

---------

Co-authored-by: Vlad Brezae <brezaevlad@gmail.com>

* [SGen/Tarjan] Handle edge case with node heaviness changing due to deduplication (#113044)

* [SGen/Tarjan] Handle edge case with node heaviness changing due to deduplication

Do early deduplication

Fix Windows build

Add test cases to sgen-bridge-pathologies

* Move test code

* Remove old code

* Add extra check (no change to functionality)

* Disable test on wasm

---------

Co-authored-by: Vlad Brezae <brezaevlad@gmail.com>
Co-authored-by: Alexander Köplinger <alex.koeplinger@outlook.com>