Propagate `Type.GetInterfaces` through dataflow analysis (#114149)
Fixes dotnet/linker#1731

We currently maintain two invariants:

1. The types returned by the API will have `.Interfaces` annotations at minimum
2. If the parent type was annotated `.All`, the types returned by the API will also be `.All`

We do this in the logic that keeps things, but we don't do this in terms of dataflow analysis warnings (the types retrieved from the array are not annotated as such, as far as the analysis is concerned). Because of the lacking annotation, we have warning suppressions in multiple places within the framework.

This fixes it.