Fix typing of BlendVariableMask condition when optimizing TernaryLogic
TernaryLogic nodes normalize small types to large types. When optimizing
a TernaryLogic node to a BlendVariableMask, if we leave the normalized
type, then the Blend uses the wrong instruction.

To fix this, change the Blend node to use the mask simd base type.
We only do this for mask base types that are small, which is
unsatisfying: there exists codegen today for double->int casts,
for example, that uses a 'double' type mask but an 'int' sized
TernaryLogic node. This works because we end up only using lane '0', as
the vector is converted to scalar. I didn't find any other case like
this, so hopefully if the mask type is small we can safely use it.

Fixes #114572