Tar: Write unused GNU tar entry size, ctime and atime bytes in the expected format (#114868)
* In the GNU tar entry format, the unused bytes of the size metadata field should be filled with "0" characters. The bug was that we were filling them with nulls like in the other formats.
Added sync and async tests to validate directly in the memory stream that the unused bytes in the size field are indeed written as "0" characters.

* In the GNU tar entry format, the atime and ctime fields (unique to this format) need to be written as null characters if their value is the Unix Epoch. The bug was that we were writing them as "0" characters as other GNU metadata fields. This was causing tools like 7-zip to show an extra unexpected folder with a name formed by just "0" characters. Other tools were unable to read these archives as they were considered corrupt.
Added sync and async tests to validate directly in the memory stream that when atime and ctime represent the Unix Epoch, they are indeed written as null characters.

* Create a separate timestamp writing method just for GNU timestamps. Regular timestamps (mtime) should not be affected by this.
Modify all the sync and async checksum shared code so that they reflect the changes in the expected values of size, atime and ctime.