Enlarge stack overflow handling stack (#114956)
This is an attempt to fix rare failures in the stack overflow tests that
I was unable to repro locally. The logs I've added so far seems to indicate
that the thread handling the exception crashes with another stack overflow
at the point when it waits for a secondary thread to finish dumping stack
trace.