Make specifying just `DllImportSearchPath.AssemblyDirectory` only search in assembly directory (#114756)
Currently, if `DllImportSearchPath.AssemblyDirectory` is the only flag specified as the search path, the runtime looks in the assembly directory first, then falls back to the OS search behaviour. This is how the flag was designed and has behaved for its existence, but has caused confusion in the past. The current behaviour is inconsistent with how flags are usually used and also leaves the developer without a simple way to specify that only the assembly is desired.

This change updates coreclr, mono, and nativeaot to treat specifying only `DllImportSearchPath.AssemblyDirectory` as only searching the assembly directory - no fallback to the OS search behaviour.