Tar: Adjust the way we write GNU longlink and longpath metadata (#114940)
* Fix 3 GNU longlink and longpath bugs

The Tar entry types Longlink and Longpath, which are exclusive to the GNU format, are metadata entries used to store longer-than-expected link or name strings (respectively). They precede their actual entry, and they are not visible to the user, only the data they hold is visible.

Most of the regular header metadata in this metadata entries is not relevant, but we need to make sure it is well formed so other tools can read them properly.

There are 3 fields that were not getting their values stored as expected by other tools:

- The mtime, ctime and atime fields need to be set to the Unix Epoch so it shows up as zeros (0x30) by default. We were storing them as null (0x0) characters instead. For consistency, if a regular entry happens to get its mtime set to the Epoch, it should also be shown as zeros (0x30).
- The uid, gid, uname and gname fields were not being set. These fields aren't really important in a metadata entry, but other tools set this to the default value (uid=gid=0 and uname=gname=root). Based on the fact that 'a value' is preferred for these fields, we can set them to the same value that the real entry has (not root).
- The long text value (either link or name) should be null terminated, and this extra character should be counted in the size of the longpath. The library libdpkg, used by the dpkg-deb tool in Debian, now expects a null terminator in longlink and longpath data starting in Debian 13. Without the terminator, these entries are considered malformed.

* Test to check mtime null values when set to UnixEpoch.

* Make sure empty uid and gid are written as '0' chars.

* Adapt existing tests that check atime and ctime so that they also check those values in the LongLink and LongPath stream bytes themselves.
Also should check uid=gid=0 and the copying of uname and gname into the LongLink/LongPath entries.
Also should confirm that the data length in LongLink and LongPath entries is counting a null terminator in the data.

* Address feedback: set to 0x30 the fields that expect octal and their value is 0 (except the last char). Adjust the new tests to take care of that.

* Update the checksum tests

* Delete prefill line.

* Bring back the if branches instead of a Math.Max. Make sure to also apply the if check >=0 to mode. Adjust tests to verify mode as well.

* No need for async MemoryStream.WriteAsync (it calls Write internally).

* Delete temp test.