Run outerloop pipeline only for release branches, not staging/preview (#114971)
I noticed we were sending 30k workitems each day to various Helix queues, this is due to the outerloop pipeline having a schedule triggere which matches old preview branches or the staging branches.

This fixes that and also sets it so that the trigger only applies if there are actual source changes.