JIT: boost inlining for methods that may return small arrays (#114806)
Look for inlinees that may be allocating and returning small fixed sized arrays.
When inlined, these array allocations may end up non-escaping and be stack allocated.

This analysis is approximate; we can't tell for sure in the IL scan what the array
size is, and we can't easily tell if the allocated array is actually returned.

Contributes to #113236