[cdac] Create transport package and rename the native cdac reader library (#114812)
This change adds new cdac transport packages that are built for Windows, MacOS, and Linux. Each package contains a host specific native cdac library that is meant to ride along with various diagnostics tools.

The native library that is part of the transport package was renamed from 'cdacreader' to 'mscordaccore_universal'.
 
Contributes to https://github.com/dotnet/runtime/issues/108720



---------

Co-authored-by: Jan Kotas <jkotas@microsoft.com>