JIT: use root compiler instance for sufficient PGO observation
During inlining, we evaluate some aspects of an inlinee's viability
while importing its direct caller. If that caller is not the inline
root we may make inconsistent observations of the overall state of
PGO. So for PGO observations always consult the root compiler.

For example, the root R may have decided to inline a small method A that
did not have PGO (say because of minimal profiling or lack of PGO for
always inlined R2R methods), and that method calls another method B;
we want to evaluate the viability of B using the PGO state of R, not of A.