[RuntimeAsync] Merge from labs, all except JIT (#114675)
This PR includes everything from the runtimelab implementation of Runtime Async feature, except: JIT changes and tests.
Some changes in between VM and JIT are hard to separate, so included here as well.

The main effect of the VM changes is classification of methods by MethodReturnKind. We recognize methods that return Task/ValueTask/Task<T>/ValueTask<T> as special.

A task-returning method definition will now get two methoddescs in the type system:

- one has the original Task-returning signature
- another is T-returning, but called with a calling convention that uses a continuation object to handle cases of incomplete result leading to suspending/resuming the chain of calls.

We call the pair "variants". For most purposes the methods are identical and parallel - i.e. if generic, both have exactly same set of type parameters, if task-returning variant overrides another one in the base class, the async variant overrides the async variant in the base, etc...
Both variants can signal that the result is not ready yet - via incomplete Task or by returning a Continuation object.

The purpose of this duality is that async variants can be much more efficient when suspension does not happen, leading to substantial savings when a chain of awaits may be turned into a chain of calls in "async space".

From the public/external point of view the async variants are an implementation detail that "does not exist", and can change in future releases. Async variants should not show up in reflection, for example.

In terms of implementation logic one variant contains the user code/IL and another is a thunk that forwards to it. Which one contains the user code depends on presence of MethodImpl.Async. If there is no MethodImpl.Async the user code belongs to the task-returning variant, otherwise to the async variant.

Note that for the most part Runtime Async feature is pay-for-play. JIT features will only engage if the user code is compiled with a compiler that is enlightened to use the new API for async/await.
However, the creation of variant pairs for task-returning methods must happen unconditionally prior to possible use. Since these are just entry points in the type system, the impact should not be high. We will, of course measure and see if some action is needed on that.

In this PR the generation of variants will require FEATURE_RUNTIME_ASYNC, which will not be defined by default, so by default none of the above will happen.

Fixes #114310