JIT: Avoid recording uncanonicalizable loops (#115223)
The example has a loop whose header is the beginning of a 'finally'
clause and whose backedge is a CALLFINALLY edge. It is not possible to
canonicalize such a loop: creating a new preheader will result in a new
beginning of the finally clause, and by IR invariants the CALLFINALLY
edge must jump to it. But that means the CALLFINALLY edge is no longer a
backedge.

The loop structure here is spurious and arises because of two-pass EH,
so it is unlikely we will lose much by just discarding these cases.