JIT: more bv-centric refactoring for escape analysis
Rework connection graph building in anticipation of having
potentially different sorts of escapes happen as we walk up a tree
and switch which resource we are tracking.

For example we may discover that a local doesn't escape in this tree,
but its fields can escape.

Make the ancestor walk responsible for recording escapes, instead
of deferring to its caller.

Also address a few small things deferred from the previous refactor.