Interpreter EH support in the runtime (#114649)
* Interpreter EH support in the runtime

This change adds support for exception handling for cases when
interpreter frames participate in the process. That means when an
exception is thrown from an interpreter frame or when it is propagated
over interpreter frames. This doesn't add all the exception handling
support to the interpreter itself, there is a follow up PR that contains
that code.

* Fix arm64 build

* Fix few issues

* Fix non-windows amd64 builds

ifdef-out setting m_SSP in the InterpreterFrame

* Fix x86, arm and arm64 build break

* Fix incorrect assert

We cannot try to get the code manager from a crawl frame when it is not
on a frameless frame.

* Resuming after catch via native exception

This commit moves the resuming after catch to using native exception
handling instead of fragile context capturing, which was not correct
anyways.

It also adds handling of exceptions comming out of native runtime
methods called from the interpreter.

* Remove some unneeded stuff and fix Unix build

* Move the saved InterpExecMethod context regs

* Move the saved registers to the StackFrameIterator
* Fix MUSL build
* Add {Get/Set}{First/Second}ArgumentRegister to all architectures and use
  those in ExecuteFunctionBelowContext.

* Few fixes

* Fix builds with disabled interpreter

* One more MUSL build fix

* Implement proper GetFuncletStartAddress for interpreter

* Fix problem with missing managed FEATURE_INTERPRETER define

* Revert "Implement proper GetFuncletStartAddress for interpreter"

This reverts commit d967861d818e98af2d5eac6c6dee5492319c4353.