Fix generic interface none-virtual method vt_slot issue.
In case of a generic interface, build_imt_slots calculate
the vt_slot to find the index in the vtable for the implementation.
Currently implementation assumed interface methods are either static
or virtual, but it's possible to have private/sealed methods that
won't end up as virtual. This is handled when building the vtable,
by not including these methods, but we didn't handle that scenario
when building the IMT slots, ending up with wrong vtable slot,
potentially outside of allocated memory.

Fixes https://github.com/dotnet/runtime/issues/113958.