Intepreter to JIT/AOT calls
This change adds support for making calls from the interpreter to JIT/AOT generated code.
For each target method, it parses the signature and creates a list of
hand written asm routines that transfer the arguments from the
interpreter stack to the CPU registers / stack based on the native
calling convention, call the target method and then places the return
value to the interpreter stack. This list is cached in the
MethodDescData so that for repeated calls to the same method, it doesn't
need to be re-generated.