Add packaging information to windowsdesktop transport package
Contributes to https://github.com/dotnet/windowsdesktop/issues/4998

Adds a props file that gets auto-imported when referencing the internal windowsdesktop
transport package. This file defines an item `PackageLibrary` that includes all the
information for the runtime assemblies that get redistributed inside the package.

This information can then be used in windowsdesktop to auto-generate parts of the
PackageOverrides.txt file.