throw an exception instead of infinite loop in sort_mark_list (#115492)
we had a customer who observed an infinite loop in sort_mark_list due to heap corruption - the object that got marked was in a region that was already freed so region_limit is 0. we detect this and throw an exception which is easier for prod diag than having to deal with an infinite loop.