JIT: Kill assertions about struct local when skipping field stores in block morph (#115556)
If we skip storing a field local because we know it is dead, then assertions
about the struct local are no longer valid, so invalidate them in that case.

Without this change we could end up using the assertion about the struct local
to remove a future store to the particular field, which would leave no zeroing
of that field, and would be observable.