Notify attached debugger of NAOT exception
Native AOT does not use C++ or SEH exceptions, meaning that debuggers do
not recognize the exception throw/catch sequence and cannot break on
first-chance exceptions. On Windows, we can use OS functionality to
throw and catch an SEH exception when a Native AOT exception is thrown
and a debugger is attached, effectively acting as a notification to
debuggers that an exception is in flight. The exception code used here
is the same one as CoreCLR, so the WinDbg command `sxe clr` would also
be sufficient to catch Native AOT exceptions.

Fixes #115514