Write barrier without any RWX pages (#114982)
Add a new config switch DOTNET_UseGCWriteBarrierCopy to control whether or not to use a copy of the WriteBarrier code instead of assembly code embedded in the coreclr binary. Typically we do this to improve performance, but this change enables a new path where we just have the barrier as assembly code in the coreclr binary that does not require any RWX behavior. Also, with this change, the 1 copy of the code which is used for this scenario is shared between NativeAOT and coreclr.

This change will change the write barrier used in some situations on CoreCLR to be the same as the write barrier used on NativeAOT.

In addition, we now have a copy of NativeAOT variant of the write barrier for Linux X86, although the other NativeAOT support is not present.

Finally, the Linux Arm32 version of the NativeAOT write barrier could now be enabled to use software write watch, like CoreCLR does. This is not enabled in this change, although the code has been written into the write barrier (as it was needed for CoreCLR).