Instrinic support for transient codegen.
The previous transient codegen paths only handled
lifetime management when the method had no
metadata back IL (that is, RVA == 0). The InstanceCalli
logic is marked as an intrinsic and has metadata,
but that metadata is copied and then modified falling
back to a middle ground that the existing logic didn't
handle properly. The tricky scenario is when these
new instrinsic types were inlined (lifetimes were
being ignored). This change fixes that up but did require
a bit of clean-up in the UnsafeJitFunction as it relates
to the JIT and interpreter interaction.