JIT: fix issue in redundant zero init opt
If there is a redundant field zero def followed by a nonzero USEASG def in a block with
EH flow, redundant zero init will improperly mark the field as not def in the block, leading
to an SSA assert when it processes the USEASG def and can't find a PHI in the handler.

Treat USEASG defs as full defs. If analysis has reached this point then all fields are zero and
so a full and partial def have the same effect. If the value being stored is not zero then
accounting for the def (which will remain) prevents the optimization from removing any fields from
the block def set.

Fixes #115123.