Interpreter to JIT/AOT calls (#115375)
* Intepreter to JIT/AOT calls

This change adds support for making calls from the interpreter to JIT/AOT generated code.
For each target method, it parses the signature and creates a list of
hand written asm routines that transfer the arguments from the
interpreter stack to the CPU registers / stack based on the native
calling convention, call the target method and then places the return
value to the interpreter stack. This list is cached in the
MethodDescData so that for repeated calls to the same method, it doesn't
need to be re-generated.

* Calling convention testing

* Cleanup, comments and apple arm64 fix

* Fix test build break and cleanup Apple arm64 stack args handling

* Fix build break

* Fix Unix x64 build break

* Fix some contracts and a bug in args by ref introduced in a previous commit

* Move to allocations from LoaderHeap

* PR feedback

* PR feedback 2

* PR feedback 3

* Wrap all the asm helpers in #ifdef FEATURE_INTERPRETER