Fix reporting of foreign native exception not handled by runtime
When a native exception occurs on a foreign thread, it is propagated
through managed frames and then flows into the foreign native frames and
is handled there, the runtime still reports the exception as unhandled
via AppDomain.CurrentDomain.UnhandledException event and also prints out
the unhandled exception message. Even though the process then continues
running just fine.

This change fixes the behavior so that only managed exceptions thrown by
the .NET runtime are reported as unhandled in such case. Foreign
exceptions, like e.g. C++ exceptions, are not reported as unhandled and
they are just propagated into the foreign native frames. If that
exception is not handled by the native frames, then C++ reports them as
unhandled.

Close #115654