Intrinsic support for transient codegen. (#115639)
The previous transient codegen paths only handled
lifetime management when the method had no
metadata back IL (that is, RVA == 0). The InstanceCalli
logic is marked as an intrinsic and has metadata,
but that metadata is copied and then modified falling
back to a middle ground that the existing logic didn't
handle properly. The tricky scenario is when these
new instrinsic types were inlined (lifetimes were
being ignored). This change fixes that up but required
clean-up in the UnsafeJitFunction code paths.
---------
Co-authored-by: Jan Kotas <jkotas@microsoft.com>