Implement CEE_LDOBJ and CEE_STOBJ in the interpreter via ldind/stind (#115900)
ldind and stind are actually aliases for ldobj and stobj according to ecma. We already have helper code for generating appropriate intop_xxind opcodes based on type, so I just reused those helpers to generate the appropriate interpreter opcodes to do the ldobj/stobj. This meant I didn't need to add new opcodes.

cpobj is not covered since as far as i can tell roslyn never generates it. The only way I found to generate it (other than a handwritten IL test) is to use pinvoke marshaling because the marshaling generator will use it. So once we do pinvoke marshaling we'll need cpobj and that's probably the right point in time to add it.